some restructuring of library to make it easier to navigate
