+++ mgpi/__init__.py
@@ -1,3 +1,10 @@
 """a framework to construct Gaussian Process interpolators for multi-dimensional tabulated data
 """
-__author__ = "Reed Essick (reed.essick@gmail.com)"
+__author__ = "Reed Essick (reed.essick@gmail.com), Ziyuan Zhang (ziyuan.z@wustl.edu)"
+
+#-------------------------------------------------
+
+from .utils import *
+from .kernels import *
+from .interpolators import *
+from .io import *
